synth_sod/pyproject.toml
[project]
name = "synth-sod"
version = "0.1.0"
description = "Synthetic dataset for salient object detection"
readme = "README.md"
authors = [{ name = "S3OD Team" }]
requires-python = ">=3.9"
dependencies = [
    "accelerate>=1.10.1",
    "albumentations>=2.0.8",
    "diffusers>=0.35.2",
    "dinov2>=0.0.1.dev0",
    "einops>=0.8.1",
    "fire>=0.7.1",
    "huggingface-hub>=0.36.0",
    "hydra-core>=1.3.2",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-core>=0.3.79",
    "matplotlib>=3.9.4",
    "numpy>=2.0.2",
    "opencv-python>=4.12.0.88",
    "peft>=0.17.1",
    "pydantic>=2.12.4",
    "pytorch-lightning>=2.5.2",
    "scikit-learn>=1.5.2",
    "tensorboard>=2.19.0",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "transformers>=4.46.2",
]

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"
